+++ queries/zig/references.scm
@@ -1,20 +1,19 @@
 ; Zig Reference Extraction Queries
 
 ; @import("module")
-(builtin_call_expression
+(build_in_call_expr
   function: (identifier) @_import
   (#eq? @_import "@import")
-  arguments: (arguments (string_literal) @ref.target_import)) @ref.type_import
+  (arguments (string_literal) @ref.target_import)) @ref.type_import
 
 ; Function calls
 (call_expression
   function: (identifier) @ref.target_call) @ref.type_call
 
-; Method calls: obj.method()
-(call_expression
-  function: (field_expression
-    member: (identifier) @ref.target_call)) @ref.type_call
-
-; Field access (struct member access)
+; Field access / Method calls
 (field_expression
-  member: (identifier) @ref.target_call) @ref.type_call
+  field: (field_identifier) @ref.target_call) @ref.type_call
+
+(field_expression
+  field: (call_expression
+    function: (identifier) @ref.target_call)) @ref.type_call